Error corrections and some claryfications
Added the API permission which is mentioned in the film below at this exact moment. 
Based on this https://youtu.be/Z9Vp2IxFzpU?t=1433

Changed the Tenant ID into client ID based on 
https://youtu.be/Z9Vp2IxFzpU?t=1807

Added a comment that the admin audit log configuration change could take up to 60 minutes to take effect.